requirements.txt
streamlit
langchain-core
langchain-openai
openai
python-dotenv
